separated generateFiber into generateCurve so I can move fibers. some new stuff in env/index.js for warping. nonfunctional parts commented out
